Adjusting useApi code, testing sign in, signup, creating shared component EmptyPanel
