+++ src/components/home/NewArraival.js
@@ -91,7 +91,6 @@ const ButtonText = styled(SharedTextStyles)`
 `;
 
 function NewArrivals({ isMainComponent, cards }) {
-  /*   console.log(cards); */
   return (
     <ArrivalLayout>
       {isMainComponent && (
