+++ src/custom_hooks/useApi.js
@@ -1,16 +1,12 @@
 import { useState, useEffect } from "react";
 import axios from "axios";
 
-// Inside the useApi hook
-
-export function usePost(url, type) {
+export function usePost(url) {
   const [data, setData] = useState(null);
   const [error, setError] = useState(null);
   const [requestBody, setRequestBody] = useState(null);
-  console.log("url : ", url);
 
   useEffect(() => {
-    console.log(requestBody);
     async function fetchData() {
       try {
         const postResponse = await axios.post(url, requestBody);
@@ -22,7 +18,7 @@ export function usePost(url, type) {
     if (requestBody !== null) {
       fetchData();
     }
-  }, [url, type, requestBody]);
+  }, [url, requestBody]);
 
   return {
     data,
@@ -42,7 +38,6 @@ export function useGet(url) {
     const fetchData = async () => {
       try {
         const response = await axios.get(url);
-/*         console.log(response.data); */
         setData(response.data);
 
         setLoading(false);
